Light meter, celsius meter.
To show the light level on a 255 scale, press A. To see the temperature in Celsius, press B. If you want to view both data in one place, press the MicroBit logo (the gold shiny logo on the V2). When you boot it up, you may experience bugs such as the text rolling on it. If that happens, please wait. For any other issues, please refer to the "Issues" section.